+++ src/router/index.js
@@ -14,13 +14,13 @@ import ResetPassword from '@/components/Pages/ResetPassword'
 import EmailVerification from '@/components/Pages/EmailVerification'
 import CreateWiki from '@/components/Pages/CreateWiki'
 import TabSettings from '@/components/Pages/ManageWiki/TabSettings'
-import TermsOfUse from '@/components/Pages/TermsOfUse'
+import TermsOfUse from '@/components/Pages/TermsOfUse/Current.vue'
 import Privacy from '@/components/Pages/Privacy/Privacy'
 import User from '@/components/Pages/User'
 import Discovery from '@/components/Pages/Discovery/Discovery'
 import Complaint from '@/components/Pages/Complaint.vue'
 import HostingPolicy from '@/components/Pages/HostingPolicy.vue'
-import TermsOfUseUpcoming from '@/components/Pages/TermsOfUseUpcoming.vue'
+import TermsOfUseUpcoming from '@/components/Pages/TermsOfUse/Upcoming.vue'
 
 Vue.use(Router)
